+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.2.0] - 2020-04-09
+### Added
+- config for text and bg color
+- use colors from AppBar as defaults
+
 ## Fixed
 - Corrected jest and babel rc configs for Babel 7 compatibility